#3 - refactoring based on the interface
